pyproject.toml
[project]
name = "langgraph-example"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "langchain-core>=1.2.17",
    "langchain-openai>=1.1.10",
    "langgraph>=1.0.10",
    "langgraph-checkpoint-sqlite>=3.0.3",
    "langgraph-cli[inmem]>=0.4.14",
    "langchain-mcp-adapters>=0.2.1",
    "python-dotenv>=1.2.2",
    "rich>=14.3.3",
    "aiosqlite>=0.22.1",
]

[tool.uv]
package = true

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[dependency-groups]
dev = [
    "ruff>=0.15.5",
]

[tool.hatch.build.targets.wheel]
packages = ["agents", "cli.py"]

[project.scripts]
chat = "cli:main"
